I figured out what's causing the error:
Mismatched binaries and headers.
The FunctionCallee returned by module.getOrInsertFunction
was acting weird, returning garbage values.
This is because I thought I was using the newer git version of llvm
in which module.getOrInsertFunction returns a FunctionCallee,
but I'm actually using the older svn version (which I have compiled),
in which module.getOrInsertFunction returns a Constant*.
Now I have to actually compile the newer version.