Got BranchCoverageRuntime to work!
I changed the output format again, but this time it's much simpler.
An output directory is created, which contains three files:
    counts: which stores three u64s of the counts of each branch type
    single: which stores pure boolean bits
    nonSingle: which stores records of the u64 non-single value and the u32 bitIndexDiff

counts is pretty simple.  It uses pwrite() to keep writing the counts on every non-single branch.

For single, a buffer of bits (as u64s) is stored.  Whenever it's full, it's flushed to the file using write().  At the end of the program, whatever's left in the buffer is flushed and a single byte recording how many bits of the last byte are meaningful.

For nonSingle, a buffer of Records is stored.  Each record contains a u1 compressed boolean for if it's multi or infinite, a u31 bitIndexDiff, which contains the difference in bit indices between the last Record, and two u32s encoding the multi or infinite value.  (Two u32s was used instead of one u64 to avoid padding.)

I've only tested this preliminarily, but it appears to be working.

I'm going to work on a simple parser now that converts to the text format.
Or I might do the block coverage first, since that should be pretty simple (hopefully).