Wrote BlockCoveragePass.  It's really simple.
I have to number the blocks with u64s though, which makes the output file larger when there aren't that many blocks.  This could be avoided by first counting the number of the blocks and determining the smallest integer size to use.  The runtime would have to have functions for each of these types though, which could be implemented.  Maybe I'll switch to this later.