Wrote BlockCoverageRuntime; still testing it and BlockCoveragePass.
BlockCoveragePass works mostly fine now,
but sometimes the first instruction in a block is required to be the first.
This means I should place the tracing call instruction
at the earliest place possible in the block,
which may not always be the first instruction.
I'm working on that now.

Also moved a bunch of stuff into .cpp implementation files
for faster compilation where possible (i.e. non-template, non-constexpr).