I figured out a lot more about what's causing the segfault. For some reason, std::function<>::~function() is causing the segfault when _M_manager(_M_functor, _M_functor, __destroy_functor) is called, which should call _M_destruct(), but the segfault is thrown instead. This is the std::function<> for ExtensionFn, the one that loads my pass.
When the segfault occurs, the stack trace starts to print.
When backtrace()/_Unwind_Backtrace() is called,
it causes another segfault, which is not handled anymore.
This segfault is because the symbols where from the shared library,
and by now they've been unloaded, so that memory is now invalid.

I think this might be the same reason why std::function<>~function() is segfaulting.
The destructor it's trying to call is in my shared library plugin,
but by the time the destructor is called,
my plugin has already been unloaded, so that destructor function is now gone.

I'm not sure how to fix this, though.
It appears to be a problem with LLVM.
I understand they need to unload plugins to save memory,
but this specific std::function doesn't even need to be a std::function
in the first place, and could just be a normal function pointer.