I figured out a solution:
LD_PRELOAD can be used to load a library before opt/clang,
meaning it won't unload until after opt/clang finishes.
Therefore, I can preload my passes, and then the std::function<>~function() destructor won't be unloaded before it's called.
This, however, is memory inefficient;
maybe not for my small passes, but for dynamically-loaded passes in general,
because the pass plugins may be large, and loading them may be unnecessary sometimes.

My solution is to make a very small preloaded library: libpass.register.so.
This contains just a single function: llvm::pass::addGlobalExtension,
which is meant to replace PassManagerBuilder::addGlobalExtension,
which RegisterStandardPasses::RegisterStandardPasses calls.

It calls PassManagerBuilder::addGlobalExtension,
but takes a raw function pointer for the ExtensionFunc,
instead of a std::function (ExtensionFn).
Therefore, the std::function is created in libpass.register.so,
which isn't unloaded until the end,
and therefore std::function<>~function() can be called on llvm_shutdown().

While the std::function thing is still a major bug in LLVM's pass registration,
this is a very good workaround that should suffice.