Tried to integrate LD_PRELOAD into cmake as llvmPreloadPassRegister(), but it's not quite working yet.
When I set LD_PRELOAD manually for clang++
(either each time, or exporting it in the shell),
it runs, loads the passes, and tries to run them,
but there are errors.
I think this has to do with running the passes alongside other passes,
instead of in a completely separate opt invocation.
But at least the stacktrace is fully printed,
so debugging shouldn't be as hard.

Also added getPassName() to passes,
since RegisterPass<> isn't being instantiated with their names anymore.