Adding leftover changes from before.
New change: Had to remove debug headers from test builds entirely
(even though the header was never included),
since debug.cpp was getting instrumented, including libstdc++,
which caused infinite recursion in the runtime.

Most of the other changes are either removing debug statements
or are for klang.cpp, which is supposed to be a clang substitute
that also calls my pass at the right times.