Added BinaryFunctionFilter, which filters out functions from given object files using llvm-nm.
For each added object file, llvm-nm is run and
the results filtered and cached to create a set of function names to filter.
This is used by both instrumentation passes to avoid
instrumenting functions that are in the runtime itself,
to avoid any possible recursive loops.

I'm still working out the kinks of BinaryFunctionFilter,
but it works with preliminary testing right now,
both standalone and in the passes.
Now using_libstdc++.cpp compiles and runs correctly.