Added io::ReadOnlyMappedMemory and io::Popen to simplify BinaryFunctionFilter.
Successfully (I think) enabled the runtime function filtering in the passes.
They both use the same BinaryFunctionFilter that contains both runtimes.

When reading from llvm-nm in BinaryFunctionFilter,
I now first put it into a StringSet, then write it to the cache,
and then re-read the cache into the main StringSet.
I'm putting them into a set because there are repeats,
so filtering them out will make reading from the cache faster.
I'm also creating the set twice because I'm using the mmap'ed memory
from the cache file for the main StringSet, so I still need to do that.
And the cache should be used almost all the time in practice,
so it shouldn't really make a difference anyways.

Added macros for the binary paths of the runtimes in CMake
using add_compile_definitions and target file generator expressions,
so I can use them in creating the BinaryFunctionFilters for the passes.

Fixed non-standard includes (full paths, all quotes unless std library).