Added fuzz.autotools.sh, which compiles an autotools project w/ my passes.
I wrote it around ccache and it works for it,
with the resulting executable working correctly.
Now I have to test if for other programs, like the GNU binutils programs.

I also added post build custom commands in CMake
that copy the built targets to
build/lib (copyLibrary) or build/bin (copyExecutable).

BlockCoverage seems to be working perfectly,
but BranchCoverage seems to not be instrumented every branch,
and only instrumenting exactly 1 branch.