Finished writing signal::Handler, but untested.
I tried to be really careful about catching all the errors in Handler,
keeping track of the previous signal handlers and running them after mine,
with as much of the correct flags and masks as possible.
I keep the same mask and most of the flags,
only changing them on a case by case basis
when necessary for my signal handler to work,
and recreating the flag behavior in my code when possible.
But in some cases, that behavior has to come from the kernel,
so there's nothing I can do about it.

I ran ccache with strace and confirmed that it does use execve,
so that's what's causing the problems in this case.

However, I realized ccache and other programs might add signal handlers
after I register mine, and they don't always keep track of the previous one
and try to call it afterwards or something,
meaning my signal handler will just be thrown away.
I guess the only way around this is to insert calls using my passes
that re-register my signal handler.

To start doing this, I added a static reRegister function
that re-instantiates the Handler singleton instance.
This is the function that can be inserted after calls to sigaction() and signal().

To ensure that I don't add this handler as one of the sub-handlers,
I check if the sa_sigaction field == this handler and skip it if it does.
But if the program calls sigaction() and
correctly wraps the old sigaction, then this won't work.
To solve it in this case, I should have an atomic boolean / lock
that I set before my signal handler starts and release afterward.