Finished most of the threading issue.
ProcessLifeCycles is the global owner of all LifeCycles for each thread,
i.e. the ThreadLifeCycles aren't thread_local.
Instead, on thread creation, a thread creates a thread_local
ThreadLifeCyclesOwner, which owns a ThreadLifeCycles,
which itself is part of the std::list (doubly-linked list) of ThreadLifeCycles.
ThreadLifeCyclesOwner creates and adds a new ThreadLifeCycles to the list,
storing a std::list iterator so that it can delete itself in O(1) time.
This deletion is done in the destructor,
so upon normal thread death, it is called (b/c it's a thread_local).

ProcessLifeCycles still stores the list to all non-deleted ThreadLifeCycles,
so it can still delete all of them at the end when,
for example, a signal is sent to kill the process.

In ThreadLifeCycles, there are two ways of destructing.
One way, when the calling thread is the same thread as the LifeCycles,
the destructor of the LifeCycles is normally called.
But when the calling thread is different,
meaning the thread of the LifeCycles might be concurrently
writing data to internal data structures like a buffer,
we can't just call the destructor while the thread is still concurrently executing.
To fix this, we first kill that thread using tgkill with SIGKILL,
which unconditionally destroys that thread.
Then we can safely call the destructor.

In ProcessLifeCycles and ThreadLifeCyclesOwner,
when adding and removing a ThreadLifeCycles from the list,
we have to guard the list from concurrent access.
I could do this using a lock-free concurrent linked-list data structure,
but there's none in the standard library or a simple implementation I can find online,
so I'm using mutexes.

At first, I thought I couldn't use a mutex because, while adding,
a signal could be sent and in the signal handler everything destructed,
so we'd have to lock a mutex that's already been locked.
To fix this, I use a recursive mutex (in STL, but using pthreads),
combined with my AdaptiveMutex so get the fast spin lock behavior in the common case.
A recursive mutex allows the same thread to repeatedly lock the mutex,
and it's only released when the thread releases all of its locks on the mutex.