Re-implemented the threads life-cycles, correctly this time I think.
Before, I was using tgkill w/ SIGKILL to kill each thread so that
it would immediately stop executing so that on the calling thread
I could clean up all the LifeCycles.

However, tgkill isn't instantaneous, and even if it was,
sending SIGKILL, which is unstoppable, kills the whole process,
not just the thread, since signals are process-based,
and tgkill is the only way to target
(use the signal handler from) a single thread.

Instead, I need another way to as fast as possible
kill/clean up another thread.
One way of doing this is to use pthread_cancel and then pthread_join.
However, pthread_cancel is not always enabled on all threads,
and I don't want to change that and potentially modify the program's behavior.
pthread_cancel also delivers a real-time signal, which is suboptimal.
The program may also use pthread_cancel for other things,
which I don't want to disturb at all.
pthread_cancel also exits the thread normally,
calling the program's own thread_local destructors.
On a SIGKILL warning, for example, this isn't correct behavior,
because I only want to call destructors for my runtime.

Instead, I've decided to send a specific, non real-time signal.
By using the syscall rt_tgsigqueueinfo, I can send a signal
to a specific thread along with a siginfo_t.
By adding a specific value to the siginfo_t,
specifically a sigval_t or aio::signal::Value,
I can tell with a 2^64 chance if I sent the signal,
or if it was randomly sent by someone/something else with the same value.
If this magic number (which I generate randomly each time)
matches what I sent, then I can safely destructor the thread local LifeCycles.
If it doesn't match, then I just re-run the original signal handler.

For this special signal, I chose SIGILL (illegal opcode).
It's a hardware signal, meaning when it is normally generated,
it is delivered synchronously and unconditionally,
meaning it cannot be masked.
Therefore, when I prohibit masking SIGILL so that my signal handler works,
it won't affect the user's program much,
because it won't affect hardware-generated SIGILLs;
only weird user-generated SIGILLs will be affected.
Furthermore, compared to SIGSEGV, SIGFPE, and SIGBUS,
SIGILL is much more rare (from my experience).

To do this, I added ThreadList, a generalization of
the previous ProcessLifeCycles and
ThreadLifeCyclesOwner (which became ThreadData now).
This is used by the new ProcessLifeCycles,
which just maintains a list of thread IDs (tid).
Each thread owns a ThreadData for this tid,
and when it exits normally, it removes this tid from the list.

ProcessLifeCycles::destruct() goes through the list of tids
and sends a SIGILL with the magic number to each thread
(see hooks::lifecycle::signaling::Sender).
Then it waits on the condition variable (see concurrent::ConditionVariable,
a combination of std::mutex and std::condition_variable)
for when the list of tids is empty and then returns
(at which point, the caller might re-raise a death signal).

ThreadLifeCycles.{h, cpp} is a lot simpler now.
In fact, it's not even its own separate class,
but just a thread_local LifeCycles with an onCreation() for ODR use
and an onDestruction() that explicitly deletes the LifeCycles.

These thread::onCreation() and thread::onDestruction()
are called by onThreadCreation() and onThreadDestruction()
in ProcessLifeCycles, which just also ODR uses and deletes the owning tid node.

ProcessLifeCycles also has a onProcessCreation() and onProcessDestruction(),
which just call the reconstruct() and destruct() methods.
reconstruct(), which is called on fork() for example,
and just clears the tids and re-assigns the thread_local own.
This is because on a fork(), only the calling thread is duplicated.

I also added hooks::lifecycle::signaling::Receiver,
which is the other end of Sender.
It checks if a signal was indeed sent by Sender with the magic number,
and if so, it calls ProcessLifeCycles::onThreadDestruction().

I also added a bunch of test programs in src/misc
just for testing out the behavior of some functions.