Re-organized the actual hook implementations in hooks.cpp to separate files in hooksImpl/.
Added support for sigprocmask and pthread_sigmask
that prevents use of SIGILL (it's also in sigaction).

Added support for the outdated System V signal API: sigset and sighold.

Tried support for the outdated BSD signal API:
sigvec, sigblock, sigsetmask, siggetmask.
However, these use int as a signal mask, which is too small,
so they don't even work correctly,
so in their hooks, we just assert(false) with an error message.