Made handler::Passive re-raise the signal after handling if it wasn't the special signal.
Also tested in misc/signalOrder.cpp
if the order I raise signals is always the order I receive them,
i.e. if I send a signal and then raise SIGKILL afterward,
the original signal should be received first.

I found that this is true, but I need to unblock the signal first inside of the signal handler.
This can easily be done by using SA_RESETHAND, which unregisters the signal handler,
and SA_NODEFER, which doesn't block the signal in it's own signal handler.