Added pass.register.preload, a wrapper to its name that adds LD_PRELOAD=.../libpass.register.preload.so. It execvpe's the program's current name, searching for it in the path, and adding the LD_PRELOAD to the environment. By copying or linking it to clang-9 or opt-9, it transparently handles all the LD_PRELOAD stuff, so calling it from CMake is much easier.
I got everything to compile in Release mode.
I didn't realize assert and llvm_unreachable are disabled when NDEBUG is defined,
so I added hardAssert.h, which is just a copy of assert.h,
except that it is only enabled when NDEBUG is defined,
and it uses the name hardAssert and hardAssertPerror.