ar2bc.sh doesn't work because files from different directories can have the same name, so they aren't extracted.  To fix this, I wrote tool.ar, which runs both llvm-ar and llmv-link.
Override llvm-ar in CMake seems impossible,
so I wrote tool.ar.override to override it in rules.ninja.
Now fuzz.autotools.sh works again, but it still gets stuck
on the condition variable for no threads left.

I also combined the two runtimes into one executable.
Most of their code is shared so this makes sense,
and they're statically linked, so dead code elimination can apply.
The runtimes are only initialized the first time
their APIs (e.x. onBlock) are called,
so if the pass for a runtime wasn't run, the runtime won't be used.
Once destructed, however, the runtimes are never re-created, though.
This is done through LazilyConstructed,
which sets the pointer to 1 if it's already destructed,
and 0 (nullptr) if it wasn't constructed yet.