It works! It works successfully on all of GNU binutils.
Their autotools is set up a bit differently since there are multiple targets,
but other than that, which I worked around,
the passes and the instrumentation worked on the first try.

I haven't thoroughly tested all of them yet,
but I was only expecting to test one and they all worked.

I ran nm-new on itself and piped it through cxxfilt and it all worked.