Made the hook implementations faster.
Instead of using local statics to load the function using dlsym,
which uses a lock to prevent double initialization,
and also has to do a comparison every time,
I now just call a function pointer.

This function pointer initially points to a function that
sets the function pointer to its actual function,
so when called in the future (unconditionally),
it will just call the actual function
without having to check if it's been initialized yet.