Finished tracing true indirect calls (can't reduce to finite branch like select or virtual functions).
onInfiniteBranch can lookup the symbol for the function address using dladdr.
This symbol can be saved and used in the branch execution program,
loaded using dlsym or something similar.

Sometimes the function address could be in a non-traced function,
i.e. it's in a shared library,
so we instead need to find the next traced function where execution resumes.

To do this, we call onFunctionStart at the start of every function.
Usually this function does nothing, but if onInfiniteBranch
determined we are entering un-traced code, a flag is flipped
to indicate to onFunctionStart it should record the function's address
using onInfiniteBranch in the same way.