I figured out the best way to integrate BranchCoverage and BranchExecution.
onFunction(u64 functionIndex) is called at the start of every function.
Normally, this does nothing, except right after an indirect call.

Before an indirect call, onInfiniteBranch() is called,
which sets a flag that we're in un-traced code currently.
If this flag is true in onFunction, the function index is recorded,
and the flag is reset.
This works even if the indirect call is to a dynamically linked function.

Each function is given an index (like in BlockCoverage),
and this index will stay consistent between BranchCoverage and BranchExecution.
In BranchExecution, we'll have a function table of every function
indexed by these function indices, so we can quickly lookup any function.

The number of functions is also recorded at compile time,
so an indirect call is in effect a multi branch (the index / numFunctions).
Therefore, onFunction() directly calls onMultiBranch().

Furthermore, I re-added the numBranches and numCases arguments
to onMultiBranch() and onSwitchCase().
Since during BranchExecution, we know the numBranches,
we can do even better than LEB128 varint encoding,
we can directly encode the significant bits into the BitWriteBuffer,
where single branches are written to.

This means we can store all the branch information in a single branches.bin file.
Its order is arbitrary; only the program following these branches can decode it,
and this is exactly what is done during BranchExecution.