Split up runtimes because dead code elimination wasn't working.
Fixed a stupid bug in parse.blocks.  It now accepts a filename to parse.

Fixed the source map filename finding.

edges.bin is sometimes bigger than blocks.bin,
even though blocks.bin is recording more information.
This is because edges records a block twice when it's a back and front edge.
There's nothing I can really do about this, though,
and the delta LEB128 encoding is already optimized for repeats.