Made a lot of progress in BranchExecutionPass fixing bugs, but it's still buggy.
I'm trying to just provide undef values for all the functions arguments
and keep the function types the same,
since changing the function call wasn't working.
It works mostly, but crashes on compiling ada_demangle for some reason.

Calling any generic function as void() might not work, though.
All the functions should be converted to void()
once dead argument elimination has ran, but until then, they calls are UB.
If the return types are all void, this should be okay, but they aren't.
I can try to change the return types to void
(since it doesn't change the number of arguments, which is the hard part).

Or I could just create new copies of each function as void().